removed column check on User
User was checking for a 16 length tuple to represent the postgres User schema. Once changed, the code now failed. Added a conditional check to the dateofbirth column which can either be returned as NULL or a value from postgres. If the value is recieved we will populate the User model with a STR->Datetime Object