Fixed AddUser Employee Options
In previous versions of create-user the name of a field such as is-admin, is-dentist, or is-admin would only show up if the option was selected so it was enough to look for the existence in the form pushed to the flask POST HTTP. It seems like this was changed ~ the field was always pushed on POST with 0/1 as false/true (resp.). I've kept the check for the field in-case a malformed HTTP POST is sent to the server but it will also check the value passed (0, 1, etc. ~ see above). Tis is Fixed gn