fixed create_employee
the checkboxes were not getting sent over POST, added an invisible input that gets a value of 'on' if the box is checked and 'off' if not checked, done 1 woo hoo. still no work - sad hours. fix 2 ; == 1 check for existenence needs to change to == on on lines 224, 237, etc. ~ we fixed ~ party time